Fix #635, Reduce use of uint32, add more OSAL typedefs

+++ src/bsp/generic-linux/src/bsp_console.c
@@ -72,7 +72,7 @@ static void OS_BSP_ExecTput(const char *cap, const char *param)
    OS_BSP_ConsoleOutput_Impl
    See full description in header
  ------------------------------------------------------------------*/
-void OS_BSP_ConsoleOutput_Impl(const char *Str, uint32 DataLen)
+void OS_BSP_ConsoleOutput_Impl(const char *Str, size_t DataLen)
 {
     ssize_t WriteLen;
