+++ src/os/portable/os-impl-no-network.c
@@ -56,7 +56,7 @@ int32 OS_NetworkGetID_Impl(int32 *IdBuf)
  *           See prototype for argument/return detail
  *
  *-----------------------------------------------------------------*/
-int32 OS_NetworkGetHostName_Impl(char *host_name, uint32 name_len)
+int32 OS_NetworkGetHostName_Impl(char *host_name, size_t name_len)
 {
     return OS_ERR_NOT_IMPLEMENTED;
 } /* end OS_NetworkGetHostName_Impl */
